During kitchen sink Android build (from react-native cli), an error was thrown: resource drawable/ic_launcher_background (aka com.kitchensink:drawable/ic_launcher_background) not found.
I attempted ./gradlew clean and problem was not on my end. If adaptive icon configuration was done through android studio previously and a file was excluded (or external) to the project, that could be the source of the error.

I rolled this back so kitchen sink android will build properly. I deleted the folder: ant-design-mobile-rn/rn-kitchen-sink/android/app/src/main/res/mipmap-anydpi-v26/

This is a very minor fix, not much is affected.

Currently, both antd icons with the android and ios kitchen sink are displayed properly after the build.